requirements.txt
# Pipeline: ControlNet + Img2Img + Where's Waldo merged checkpoint

torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
huggingface_hub>=0.20.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
# Pipeline C: SoftEdge control maps (PidiNet / Annotators)
controlnet_aux>=0.0.7
# Optional: YOLO person detection (other pipeline stages)
ultralytics>=8.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
